Project Spartan Pre-Alpha 1.0 build 2
Added a scenario switcher, player sprites (WILL BE CHANGED) and testing out moving. Moving is currently not working and I am working on a fix to it. Will most likely not come soon as I am experimenting with the new amazon lumberyard engine. I also reorganized a bit of code.